+++ app/src/main/java/com/example/cupcake/fragments/FlavorFragment.kt
@@ -13,9 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.example.cupcake
+package com.example.cupcake.fragments
 
-import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -23,19 +22,19 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.navigation.fragment.findNavController
-import com.example.cupcake.databinding.FragmentSummaryBinding
+import com.example.cupcake.R
+import com.example.cupcake.databinding.FragmentFlavorBinding
 import com.example.cupcake.model.OrderViewModel
 
 /**
- * [SummaryFragment] contains a summary of the order details with a button to share the order
- * via another app.
+ * [FlavorFragment] allows a user to choose a cupcake flavor for the order.
  */
-class SummaryFragment : Fragment() {
+class FlavorFragment : Fragment() {
 
-    // Binding object instance corresponding to the fragment_summary.xml layout
+    // Binding object instance corresponding to the fragment_flavor.xml layout
     // This property is non-null between the onCreateView() and onDestroyView() lifecycle callbacks,
     // when the view hierarchy is attached to the fragment.
-    private var binding: FragmentSummaryBinding? = null
+    private var binding: FragmentFlavorBinding? = null
 
     // Use the 'by activityViewModels()' Kotlin property delegate from the fragment-ktx artifact
     private val sharedViewModel: OrderViewModel by activityViewModels()
@@ -44,7 +43,7 @@ class SummaryFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        val fragmentBinding = FragmentSummaryBinding.inflate(inflater, container, false)
+        val fragmentBinding = FragmentFlavorBinding.inflate(inflater, container, false)
         binding = fragmentBinding
         return fragmentBinding.root
     }
@@ -53,38 +52,22 @@ class SummaryFragment : Fragment() {
         super.onViewCreated(view, savedInstanceState)
 
         binding?.apply {
+            // Specify the fragment as the lifecycle owner
             lifecycleOwner = viewLifecycleOwner
+
+            // Assign the view model to a property in the binding class
             viewModel = sharedViewModel
-            summaryFragment = this@SummaryFragment
+
+            // Assign the fragment
+            flavorFragment = this@FlavorFragment
         }
     }
 
     /**
-     * Submit the order by sharing out the order details to another app via an implicit intent.
+     * Navigate to the next screen to choose pickup date.
      */
-    fun sendOrder() {
-        // Construct the order summary text with information from the view model
-        val numberOfCupcakes = sharedViewModel.quantity.value ?: 0
-        val orderSummary = getString(
-            R.string.order_details,
-            resources.getQuantityString(R.plurals.cupcakes, numberOfCupcakes, numberOfCupcakes),
-            sharedViewModel.flavor.value.toString(),
-            sharedViewModel.date.value.toString(),
-            sharedViewModel.price.value.toString()
-        )
-
-        // Create an ACTION_SEND implicit intent with order details in the intent extras
-        val intent = Intent(Intent.ACTION_SEND)
-            .setType("text/plain")
-            .putExtra(Intent.EXTRA_SUBJECT, getString(R.string.new_cupcake_order))
-            .putExtra(Intent.EXTRA_TEXT, orderSummary)
-
-        // Check if there's an app that can handle this intent before launching it
-        if (activity?.packageManager?.resolveActivity(intent, 0) != null) {
-            // Start a new activity with the given intent (this may open the share dialog on a
-            // device if multiple apps can handle this intent)
-            startActivity(intent)
-        }
+    fun goToNextScreen() {
+        findNavController().navigate(R.id.action_flavorFragment_to_pickupFragment)
     }
 
     /**
@@ -95,7 +78,7 @@ class SummaryFragment : Fragment() {
         sharedViewModel.resetOrder()
 
         // Navigate back to the [StartFragment] to start over
-        findNavController().navigate(R.id.action_summaryFragment_to_startFragment)
+        findNavController().navigate(R.id.action_flavorFragment_to_startFragment)
     }
 
     /**
