+++ app/src/main/res/layout/fragment_pickup.xml
@@ -18,7 +18,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".PickupFragment">
+    tools:context=".fragments.PickupFragment">
 
     <data>
         <variable
@@ -27,7 +27,7 @@
 
         <variable
             name="pickupFragment"
-            type="com.example.cupcake.PickupFragment" />
+            type="com.example.cupcake.fragments.PickupFragment" />
     </data>
 
     <ScrollView
