+++ app/src/main/res/layout/fragment_start.xml
@@ -18,12 +18,12 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".StartFragment">
+    tools:context=".fragments.StartFragment">
 
     <data>
         <variable
             name="startFragment"
-            type="com.example.cupcake.StartFragment" />
+            type="com.example.cupcake.fragments.StartFragment" />
     </data>
 
     <ScrollView
