+++ app/src/main/res/layout/fragment_summary.xml
@@ -17,7 +17,7 @@
 <!-- Layout for order summary -->
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".SummaryFragment">
+    tools:context=".fragments.SummaryFragment">
 
     <data>
         <variable
@@ -26,7 +26,7 @@
 
         <variable
             name="summaryFragment"
-            type="com.example.cupcake.SummaryFragment" />
+            type="com.example.cupcake.fragments.SummaryFragment" />
     </data>
 
     <ScrollView
